MediMeowAI/requirements.txt
grpcio
grpcio-tools
protobuf
langchain
langchain-community
langchain-chroma
langchain-openai
sentence-transformers
requests
zhipuai